Additional conditions for use cases
Adds a non-breaking change to make use cases support multiple conditions.
This allows for more atomic profiles which better represent real-world feature requirements.
One such requirement could be that the BMC must expose the temperature of its host CPU.
This can be modeled easily with a multi-conditional Sensor UseCase with the following conditions:
- PhysicalContext = CPU
- ReadingType = Temperature
This UseCase can then carry the specific requirements for such a sensor (update frequency, precision...).

A minimal profile containing such an atomic use case can then be used to represent the implementation requirements of a feature, and be shared between different products requirements (through inclusion with RequiredProfiles or RequiredResourceProfile).

Currently, a combination of UseCase + ConditionalRequirements can emulate the same feature, but is not practical:
- ConditionalRequirements have to be repeated for each property requirement (which there may be many of, e.g. thresholds in the case of a Sensor)
- It is impossible to subordinate a requirement to more than two conditions (one from the UseCase and one from the ConditionalRequirement)
- Readability is strongly impaired, since the conditions are split between the UseCase section and the ConditionalRequirements.

Note that while this is originally a vendor idea (Atos/Eviden firmware architecture team), interest for this change has been confirmed with client domain experts (Criteo hardware management team) at Prem'Day 2025 in Paris.